dtcalc: resolve 13 compiler warnings on 64 bit builds.

+++ cde/programs/dtcalc/help.c
@@ -154,9 +154,9 @@ void
 HelpRequestCB(Widget widget, caddr_t client_data, caddr_t call_data)
 {
   char  *helpVolume, *locationId;
-  int   topic;
+  intptr_t   topic;
 
-  topic = ((int) client_data) & 0xFFFF;
+  topic = ((intptr_t) client_data) & 0xFFFF;
   helpVolume = HELP_VOLUME;
 
   if(topic < 56)
